rename arduino to serial
